Move netcdf functions to separate module
